+++ SPOVersionManagement.psm1
@@ -1,4 +1,4 @@
-#Requires -Modules Microsoft.Online.SharePoint.PowerShell
+﻿#Requires -Modules Microsoft.Online.SharePoint.PowerShell
 
 <#
 .SYNOPSIS